Simplify directive middleware
